Rework Stitch<->Core communications to be fully bi-directional and asynchronous
Messages from the stitch are read in a separate thread and passed up to the messagebus via events
Messages from the Core are read in a separate thread and passed out using a BlockingCollection
Allow the stitch to send un-requested messages such as data and logs.
Logs are supported and are passed to the LoggingModule. Data events are noop for now
Change the StitchMessageManager to have a better API and not require a processor/callback. Example in StitchStart test program
Have a thread to check the status of the Core process PID and close the Stitch application when the core dies.
Misc cleanups and fixes, TODO notes.