Implement NodeStatus broadcast across cluster nodes
Bump Acquaintance dependency to 1.0.0-rc2 to get a fix for weak references
Fix MessageEnvelope and related logic to reliably send, receive, and publish events
Abstract JSON-handling into a utility class, for dependency isolation and consistent settings
Change the MessageTimerModule to use a registered waithandle instead of AcquaintanceTimer. This is a temporary change and might be reverted because the problem I was debugging is gone.
Misc cleanups and method extractions.