[Stitches] Enable graceful exit in stitches
Add an _exit message channel. A Stitch can choose to receive messages on this channel.
If the stitch does not, the loss of the core will just terminate the stitch
If the stitch does, it will receive the message and can respond to it gracefully and terminate at it's own pace.