[Master,Backplane] CommandJob and end-to-end commands and receipts
Cleanups, fixes and validations for backplane, envelope sending, commands, misc.
Fill in the from nodeId/networkNodeId fields in the backplane
Start with commands to start/stop and remove a stitch and a basic Ping command
Fixes to property route, validate and send outgoing messages.
Update ClusterTest test programs to send pings from the Master to the Client on each heartbeat and update job status