[Stitches] README for CrossStitch.Stitch and Cleanups for Stitch-side API
Handle _heartbeat and _exit messages in the GetNextMessage method, which provides much better results
If the Stitch is processing messages synchronously, we know that the exit or sync will only happen at a good boundary
In the case of Heartbeat/sync, we know that the application is responsive, and not just the reader thread.
Some additional cleanups in support of better documentation
README for how to create ProcessV1 stitches using the Stitch library.