[Stitches,Master] Introduce the concept of a 'FullId' for Stitches
Stitch IDs are guaranteed to be unique per node but not per cluster. To allow ID-based addressing to a stitch, we need to have full IDs.
FullIds are {NodeID}:{StitchID}
Add a StitchFullId class to handle the various formats and parsings.
Update Master to properly address messages with IDs in {StitchId} and {NodeId}:{StitchId} formats
Master handles StitchDataMessage and CommandRequest messages with IDs in these formats.