[Stitches,Master,Backplane] Package upload and stitch create over the cluster
FileTransferRequest to send package files across the network (simple implementation does not use streams or chunks yet)
Add InternalObject payload type for ClusterMessage, so we can send messages directly to the backplane and not immediately be routed to the bus
Break CreateInstanceRequest into cluster and Local- varieties, with necessary fields for each
MasterModule handles Package upload and stitch create commands, distributes packages across the cluster, and can create N instances in a single command
Update StitchModule to allow creating N instances locally at once
Add events for when Jobs are completed, so we can use this for basic orchestration and other things in the future.
Fix StitchesModule to not always need to unzip the package when we create the stitch instance (built-ins, and tests where the stitch code is already available)
Add StitchIntegration to test orchestrated command sends across a cluster
Update StitchStart test program to exercise more of the create/start workflow for stitches.
Cleanup some config stuff
Needs cleanup, reorg, tests